drm hisi use kHz for clock computations
1) Enhance all the computations to use ps period resolution and kHz frequency resolution, improving accuracy.

2) Automatically use 4 lanes when clock rate > 80MHz, for compatibility with
1080p

3) I found enable and disable of the DSI LP mode when other transmissions can
occur was necessary to make each mode work.  By trial and error I found what
is needed for each mode and implemented heuristics.

4) Some modes require one more byte lane clock compared to the calculated
amount, I added a heuristic to do it

With these changes 1080p50 and 60 modes work on all 3 of my monitors, and
720p60 can also work.  720p50 worked on one monitor but on another he was
flicking between showing it correctly and feeling the mode was invalid,
so I disabled it.

There is a colour swap problem on 1080p60 I noticed sometimes.  On some
boots what should be R, G, B are misordered as B, R, G,  I think it means we
slipped a colour byte

        sent by DSI:  RGBRGBRGBRGB
received by ADV7533:   RGBRGBRGBRGB

This happens between DSI and ADV7533 because

 - it's the first place the pixels are serialized

 - swapping monitors after this happens it never changes afterwards (ie,
   problem is not happening in the monitor receiver)

I only just noticed it so not sure what to do about it yet.

Signed-off-by: Andy Green <andy.green@linaro.org>