drm adv7511 make num lanes decideable dynamically
The existing code seems to expect there is only one setting for
the number of lanes in use.  On a pluggable HDMI device, the
number of lanes needed changes dynamically depending on the mode
that is chosen.

This patch makes the number of lanes optional in the dts, if left
out then it's decided dynamically, on the heuristic that
pixel clock > 80MHz == 4 lanes, otherwise 3

Signed-off-by: Andy Green <andy.green@linaro.org>